sidecar/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
torch==2.5.*
torchvision==0.20.*
transformers==4.47.*
lpips==0.1.*
scikit-image==0.24.*
Pillow==11.*
numpy==2.*
python-multipart==0.0.*
